pyproject.toml
[tool.poetry]
name = "exep_tools"
version = "0.0.1"
description = "-"
authors = ["yakov <yakov@canway.net>"]
homepage = "https://code.cwoa.net/rd-fy23-canway-aiops-rdc/exep-tools"
packages = [{ include = "exep_tools" }]
include = [
    "exep_tools/*.so",
    "exep_tools/**/*.so",
    "exep_tools/**/*.pyd",
    "exep_tools/*.pyd",
    "exep_tools/*/*.dll",
    "exep_tools/**/*.dll",
]
exclude = [
    "exep_tools/*.pyc",
    "exep_tools/*.py",
    "exep_tools/*.c",
    "exep_tools/__pycache__",
    "exep_tools/**/*.pyc",
    "exep_tools/**/*.py",
    "exep_tools/**/*.c",
    "exep_tools/**/__pycache__",
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
click = "^8.1.8"
python-dotenv = "^1.1.0"
pycryptodome = "^3.22.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
ipython = "^9.2.0"
requests-mock = "^1.12.1"
cython = "^3.1.0"
pytest-cov = "^6.1.1"
ast-grep-py = "^0.37.0"

[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0", "Cython"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.build]
generate-setup-file = false
script = "bin/build.py"

[tool.poetry.scripts]
exep-tools = "exep_tools.main:cli"

[tool.mypy]
files = ["exep_tools"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true



[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
